Remove seen from TypeSizeAccumulator
This fixes #15692 and does not seem to break an existing compilation
tests. The problem with seen when it comes #15692 is that seen means
that type that has repeated types will get a lower size and this
incorrectly triggers the divergence check since some of the steps
involved less repeated types.

The seen logic was introduced in #6329 and the motivation was to deal
with F-bounds. Since not tests fail it not clear if this logic is still
needed to deal with F-bounds? If it is still needed we can add a test
and instead of removing the seen logic we can make it track only types
the appear as a bound and could cause infinite recursion instead of
tracking all.

[Cherry-picked c2ef180d351c7e1e759b9199c6a6df73cdbba584]